Changed Constraint layout to linear layout in editor and add scholly activity

+++ app/src/main/java/com/example/scholarshiptracker/adapters/ScholarshipAdapter.java
@@ -83,7 +83,7 @@ public class ScholarshipAdapter extends ListAdapter<Scholarship, ScholarshipAdap
 
         }
 
-        public void bind(Scholarship currentScholarship, String name, int amount, String appliedDate, ScholarshipAdapter.onClickInterface onClickInterface) {
+        public void bind(Scholarship currentScholarship, String name, double amount, String appliedDate, ScholarshipAdapter.onClickInterface onClickInterface) {
             scholarshipName.setText(name);
             scholarshipAmount.setText(String.valueOf(amount));
             dateApplied.setText(appliedDate);
